move functions only used in one place from common file to specific
