+++ mapper/census.go
@@ -17,7 +17,6 @@ import (
 	"github.com/ONSdigital/dp-frontend-dataset-controller/helpers"
 	sharedModel "github.com/ONSdigital/dp-frontend-dataset-controller/model"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/model/datasetLandingPageCensus"
-	"github.com/ONSdigital/dp-renderer/helper"
 	coreModel "github.com/ONSdigital/dp-renderer/model"
 )
 
@@ -101,97 +100,6 @@ func populateCollapsible(Dimensions []dataset.VersionDimension, isFilterOutput b
 	}
 
 	return collapsibleContentItems
-}
-
-func mapCensusOptionsToDimensions(dims []dataset.VersionDimension, opts []dataset.Options, queryStrValues []string, path, lang string, isFlex, isMultivariate bool) ([]sharedModel.Dimension, []datasetLandingPageCensus.Panel) {
-	dimensions := []sharedModel.Dimension{}
-	qs := []datasetLandingPageCensus.Panel{}
-	for _, opt := range opts {
-		var pDim sharedModel.Dimension
-
-		for _, dimension := range dims {
-			if dimension.Name == opt.Items[0].DimensionID {
-				pDim.Name = dimension.Name
-				pDim.Description = dimension.Description
-				pDim.IsAreaType = helpers.IsBoolPtr(dimension.IsAreaType)
-				pDim.ShowChange = pDim.IsAreaType && isFlex || isMultivariate
-				pDim.Title = cleanDimensionLabel(dimension.Label)
-				pDim.ID = dimension.ID
-				if dimension.QualityStatementText != "" && dimension.QualityStatementURL != "" {
-					qs = append(qs, datasetLandingPageCensus.Panel{
-						Body:       fmt.Sprintf("<p>%s</p>%s", dimension.QualityStatementText, helper.Localise("QualityNoticeReadMore", lang, 1, dimension.QualityStatementURL)),
-						CssClasses: []string{"ons-u-mt-no"},
-					})
-				}
-			}
-		}
-
-		pDim.TotalItems = opt.TotalCount
-		midFloor, midCeiling := getTruncationMidRange(opt.TotalCount)
-
-		var displayedOptions []dataset.Option
-		if pDim.TotalItems > 9 && !helpers.HasStringInSlice(pDim.ID, queryStrValues) {
-			displayedOptions = opt.Items[:3]
-			displayedOptions = append(displayedOptions, opt.Items[midFloor:midCeiling]...)
-			displayedOptions = append(displayedOptions, opt.Items[len(opt.Items)-3:]...)
-			pDim.IsTruncated = true
-		} else {
-			displayedOptions = opt.Items
-		}
-
-		for _, opt := range displayedOptions {
-			pDim.Values = append(pDim.Values, opt.Label)
-		}
-
-		q := url.Values{}
-		if pDim.IsTruncated {
-			q.Add(queryStrKey, pDim.ID)
-		}
-		pDim.TruncateLink = generateTruncatePath(path, pDim.ID, q)
-		dimensions = append(dimensions, pDim)
-	}
-	return dimensions, qs
-}
-
-func mapFilterOutputDims(dims []sharedModel.FilterDimension, queryStrValues []string, path string, isMultivariate bool) []sharedModel.Dimension {
-	sort.Slice(dims, func(i, j int) bool {
-		return *dims[i].IsAreaType
-	})
-	dimensions := []sharedModel.Dimension{}
-	for _, dim := range dims {
-		var isAreaType bool
-		if helpers.IsBoolPtr(dim.IsAreaType) {
-			isAreaType = true
-		}
-		pDim := sharedModel.Dimension{}
-		pDim.Title = cleanDimensionLabel(dim.Label)
-		pDim.ID = dim.ID
-		pDim.Name = dim.Name
-		pDim.IsAreaType = isAreaType
-		pDim.ShowChange = isAreaType || isMultivariate
-		pDim.TotalItems = dim.OptionsCount
-		midFloor, midCeiling := getTruncationMidRange(pDim.TotalItems)
-
-		var displayedOptions []string
-		if pDim.TotalItems > 9 && !helpers.HasStringInSlice(pDim.ID, queryStrValues) && !pDim.IsAreaType {
-			displayedOptions = dim.Options[:3]
-			displayedOptions = append(displayedOptions, dim.Options[midFloor:midCeiling]...)
-			displayedOptions = append(displayedOptions, dim.Options[len(dim.Options)-3:]...)
-			pDim.IsTruncated = true
-		} else {
-			displayedOptions = dim.Options
-		}
-
-		pDim.Values = append(pDim.Values, displayedOptions...)
-
-		q := url.Values{}
-		if pDim.IsTruncated {
-			q.Add(queryStrKey, pDim.ID)
-		}
-		pDim.TruncateLink = generateTruncatePath(path, pDim.ID, q)
-		dimensions = append(dimensions, pDim)
-	}
-	return dimensions
 }
 
 // getTruncationMidRange returns ints that can be used as the truncation mid range
