+++ mapper/census_filter_outputs.go
@@ -3,12 +3,15 @@ package mapper
 import (
 	"context"
 	"net/http"
+	"net/url"
+	"sort"
 	"strconv"
 	"strings"
 
 	"github.com/ONSdigital/dp-api-clients-go/v2/dataset"
 	"github.com/ONSdigital/dp-api-clients-go/v2/filter"
 	"github.com/ONSdigital/dp-api-clients-go/v2/zebedee"
+	"github.com/ONSdigital/dp-frontend-dataset-controller/helpers"
 	sharedModel "github.com/ONSdigital/dp-frontend-dataset-controller/model"
 	"github.com/ONSdigital/dp-frontend-dataset-controller/model/datasetLandingPageCensus"
 	coreModel "github.com/ONSdigital/dp-renderer/model"
@@ -74,6 +77,47 @@ func CreateCensusFilterOutputsPage(isEnableMultivariate bool, ctx context.Contex
 	p.PreGTMJavaScript = append(p.PreGTMJavaScript, getDataLayerJavaScript(getFilterAnalytics(fDims, hasNoAreaOptions)))
 
 	return p
+}
+
+func mapFilterOutputDims(dims []sharedModel.FilterDimension, queryStrValues []string, path string, isMultivariate bool) []sharedModel.Dimension {
+	sort.Slice(dims, func(i, j int) bool {
+		return *dims[i].IsAreaType
+	})
+	dimensions := []sharedModel.Dimension{}
+	for _, dim := range dims {
+		var isAreaType bool
+		if helpers.IsBoolPtr(dim.IsAreaType) {
+			isAreaType = true
+		}
+		pDim := sharedModel.Dimension{}
+		pDim.Title = cleanDimensionLabel(dim.Label)
+		pDim.ID = dim.ID
+		pDim.Name = dim.Name
+		pDim.IsAreaType = isAreaType
+		pDim.ShowChange = isAreaType || isMultivariate
+		pDim.TotalItems = dim.OptionsCount
+		midFloor, midCeiling := getTruncationMidRange(pDim.TotalItems)
+
+		var displayedOptions []string
+		if pDim.TotalItems > 9 && !helpers.HasStringInSlice(pDim.ID, queryStrValues) && !pDim.IsAreaType {
+			displayedOptions = dim.Options[:3]
+			displayedOptions = append(displayedOptions, dim.Options[midFloor:midCeiling]...)
+			displayedOptions = append(displayedOptions, dim.Options[len(dim.Options)-3:]...)
+			pDim.IsTruncated = true
+		} else {
+			displayedOptions = dim.Options
+		}
+
+		pDim.Values = append(pDim.Values, displayedOptions...)
+
+		q := url.Values{}
+		if pDim.IsTruncated {
+			q.Add(queryStrKey, pDim.ID)
+		}
+		pDim.TruncateLink = generateTruncatePath(path, pDim.ID, q)
+		dimensions = append(dimensions, pDim)
+	}
+	return dimensions
 }
 
 func getFilterAnalytics(filterDimensions []sharedModel.FilterDimension, defaultCoverage bool) map[string]string {
